refactor: auto-split test_sensing.py
